change: refactor and move AppBar
